Refactore resource data store
